app backup scripts: allow backup creation to keep a number of backups
The number of backups kepts is handled by the nas. So backup creation
just needed to create the current backup and could dispose of
previous ones on every run.

A situation came up, "what about creating backups for serivices that
run on the nas"? In these situations, we wouldn't be copying the
backups from a remote (which would do the pruning of old backups).
So the backup creation process can now opt into preserving previous
backups it creates.