app backup scripts: add Portainer backup handler
Portainer backup via web API.

Portainer requires a username and password to get a jwt. Then the jwt
can be used to get the raw backup data, which can then be written to
file. Due to security reasons, the backup config file permissions had
to be changed to now be world readable.